+++ particles-config/src/commonMain/kotlin/me/dvyy/particles/dsl/Simulation.kt
@@ -10,6 +10,8 @@ data class Simulation(
     val dT: Double = 0.001,
     val maxVelocity: Double = 20.0,
     val maxForce: Double = 100000.0,
+    val targetVelocity: Double = 0.0,
+    val targetVelocityStrength: Double = 0.0,
     val threeDimensions: Boolean = false,
     val passesPerFrame: Int = 100,
     val size: Size = Size(),