+++ particles-kool/src/commonMain/kotlin/me/dvyy/particles/compute/simulation/SimulationParametersStruct.kt
@@ -6,4 +6,6 @@ import de.fabmax.kool.util.Struct
 class SimulationParametersStruct : Struct("SimulationParametersStruct", MemoryLayout.Std140) {
     val maxVelocity = float1()
     val maxForce = float1()
+    val targetVelocity = float1()
+    val targetVelocityFixStrength = float1()
 }